+++ R/DW_nodelml.R
@@ -1,4 +1,4 @@
-#' Compute node-marginal likelihoods of a DAG model (internal function)
+#' Compute node-marginal likelihoods of a DAG model
 #'
 #' This function computes the log-marginal likelihood of the conditional distribution of variable \code{node}
 #' given its parents in \code{DAG} under a DAG-Wishart prior on the DAG model-parameters
@@ -11,7 +11,8 @@
 #' @param U position hyperparameter of the DAG Wishart prior
 #'
 #' @return The logarithm of the marginal likelihood of \code{node}
-#' @export
+#' @noRd
+#' @keywords internal
 DW_nodelml <- function(node, DAG, tXX, n, a, U) {
   j <- node
   pa <- pa(j, DAG)
@@ -31,9 +32,9 @@ DW_nodelml <- function(node, DAG, tXX, n, a, U) {
 
   } else {
     U_paj.j <- U[pa,j]
-    U_jj <- U[j,j] - t(U_paj.j)%*%solve(U[pa,pa])%*%U_paj.j
+    U_jj <- U[j,j] - t(U_paj.j)%*%chol2inv(chol(U[pa,pa]))%*%U_paj.j
     Upost_paj.j <- Upost[pa,j]
-    Upost_jj <- Upost[j,j] - t(Upost_paj.j)%*%solve(Upost[pa,pa])%*%Upost_paj.j
+    Upost_jj <- Upost[j,j] - t(Upost_paj.j)%*%chol2inv(chol(Upost[pa,pa]))%*%Upost_paj.j
 
     prior.normcost <- -lgamma(a.star/2) + a.star/2*log(U_jj/2) + 0.5*log(det(as.matrix(U[pa,pa])))
     post.normcost <- -lgamma(a.star/2 + n/2) + (a.star/2 + n/2)*log(Upost_jj/2) + 0.5*log(det(as.matrix(Upost[pa,pa])))