+++ R/rnodeDAGWishart.R
@@ -10,6 +10,11 @@
 #' @param aj common shape hyperparameter of the compatible DAG-Wishart, \eqn{a > q - 1}
 #' @param U position hyperparameter of the compatible DAG-Wishart, a \eqn{(q, q)} s.p.d. matrix
 #'
+#' @return A sample of size one from the local distribution of \code{node} in a compatible DAG-Wishart distribution.
+#'
+#' @noRd
+#' @keywords internal
+#'
 #' @return A list with two elements; a vector with one draw for the (vector) regression coefficient and a scalar with one draw for the conditional variance
 rnodeDAGWishart <- function(node, DAG, aj, U) {
   q <- ncol(data)
@@ -25,7 +30,7 @@ rnodeDAGWishart <- function(node, DAG, aj, U) {
     out$sigmaj <- stats::rgamma(1, shape = aj/2, rate = U_jj/2)^-1
   } else {
     U_paj.j <- U[pa,j]
-    invU_papa <- solve(U[pa,pa])
+    invU_papa <- chol2inv(chol(U[pa,pa]))
     U_jj <- U[j,j] - t(U_paj.j)%*%invU_papa%*%U_paj.j
 
     out$sigmaj <- stats::rgamma(1, shape = aj/2, rate = U_jj/2)^-1